migrate from wget to echo mechanism
